pyproject.toml
[project]
name = "celine-pipelines"
version = "0.1.0"
description = "CELINE pipelines toolkit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celine-utils>=1.4.2",
]

[dependency-groups]
dev = [
    "semver>=3.0.4",
]

